simplify _BaseObject, use registry
