+++ fastkml/base.py
@@ -221,103 +221,3 @@ class _XMLObject:
                 config.etree.fromstring(string),  # type: ignore[attr-defined]
             ),
         )
-
-
-class _BaseObject(_XMLObject):
-    """
-    Base class for all KML objects.
-
-    This is an abstract base class and cannot be used directly in a
-    KML file. It provides the id attribute, which allows unique
-    identification of a KML element, and the targetId attribute,
-    which is used to reference objects that have already been loaded into
-    Google Earth. The id attribute must be assigned if the <Update>
-    mechanism is to be used.
-    """
-
-    _default_ns = config.KMLNS
-
-    id = None
-    target_id = None
-
-    def __init__(
-        self,
-        ns: Optional[str] = None,
-        name_spaces: Optional[Dict[str, str]] = None,
-        id: Optional[str] = None,
-        target_id: Optional[str] = None,
-        **kwargs: Any,
-    ) -> None:
-        """Initialize the KML Object."""
-        super().__init__(ns=ns, name_spaces=name_spaces, **kwargs)
-        self.id = id
-        self.target_id = target_id
-
-    def __repr__(self) -> str:
-        """Create a string (c)representation for _BaseObject."""
-        return (
-            f"{self.__class__.__module__}.{self.__class__.__name__}("
-            f"ns={self.ns!r}, "
-            f"name_spaces={self.name_spaces!r}, "
-            f"id={self.id!r}, "
-            f"target_id={self.target_id!r}, "
-            f"**kwargs={self._get_splat()!r},"
-            ")"
-        )
-
-    def __eq__(self, other: object) -> bool:
-        """Return True if the two objects are equal."""
-        try:
-            assert isinstance(other, type(self))
-            return (
-                super().__eq__(other)
-                and self.id == other.id
-                and self.target_id == other.target_id
-            )
-        except AssertionError:
-            return False
-
-    def etree_element(
-        self,
-        precision: Optional[int] = None,
-        verbosity: Verbosity = Verbosity.normal,
-    ) -> Element:
-        """Return the KML Object as an Element."""
-        element = super().etree_element(precision=precision, verbosity=verbosity)
-        if self.id:
-            element.set("id", self.id)
-        if self.target_id:
-            element.set("targetId", self.target_id)
-        return element
-
-    @classmethod
-    def _get_id(cls, element: Element, strict: bool) -> str:
-        return element.get("id") or ""
-
-    @classmethod
-    def _get_target_id(cls, element: Element, strict: bool) -> str:
-        return element.get("targetId") or ""
-
-    @classmethod
-    def _get_kwargs(
-        cls,
-        *,
-        ns: str,
-        name_spaces: Optional[Dict[str, str]] = None,
-        element: Element,
-        strict: bool,
-    ) -> Dict[str, Any]:
-        """Get the keyword arguments to build the object from an element."""
-        kwargs = super()._get_kwargs(
-            ns=ns,
-            name_spaces=name_spaces,
-            element=element,
-            strict=strict,
-        )
-        kwargs.update(
-            {
-                "id": cls._get_id(element=element, strict=strict),
-                "target_id": cls._get_target_id(element=element, strict=strict),
-            },
-        )
-        return kwargs
